+++ src/cmd_supprimer.cpp
@@ -7,7 +7,6 @@ CmdSupprimer::CmdSupprimer(std::vector<IElement*>& container,
       element_(elem),
       oldIndex_(0)
 {
-    // On sauve la position initiale
     auto it = std::find(container_.begin(), container_.end(), element_);
     if (it != container_.end())
         oldIndex_ = std::distance(container_.begin(), it);
@@ -15,7 +14,6 @@ CmdSupprimer::CmdSupprimer(std::vector<IElement*>& container,
 
 void CmdSupprimer::executer()
 {
-    // retirer l’élément
     container_.erase(
         std::remove(container_.begin(), container_.end(), element_),
         container_.end()
@@ -24,9 +22,8 @@ void CmdSupprimer::executer()
 
 void CmdSupprimer::annuler()
 {
-    // remettre à sa position initiale
-    if (oldIndex_ > container_.size())
-        oldIndex_ = container_.size();
+    if (oldIndex_ > (int)container_.size())
+        oldIndex_ = (int)container_.size();
 
     container_.insert(container_.begin() + oldIndex_, element_);
 }
